Implementation of complex aggregates in PostgreSQL
This project  aims at implementing complex aggregates in PostgreSQL
while computing the spatial distance histogram of a set of 3D points.
Given the coordinates of the N number of 3d particles in a box, and a
user-defined distance, we calculate the number of particle to particle
distances falling into separate categories (buckets) based on the series
of distance ranges between the particles. The expected result of the
current work is to achieve a histogram consisting the number of
particle-to-particle distances falling into a series of ranges of the
bucket width.